refactor: move composite/signals.rs to retrieval/signals.rs; delete empty composite/ module; demote pub -> pub(crate)
