--- crates/origin-core/src/composite/mod.rs
@@ -1,1 +1,0 @@
-pub mod signals;
