+++ crates/origin-core/src/lib.rs
@@ -15,7 +15,6 @@ pub mod cache;
 pub mod chat_import;
 pub mod chunker;
 pub mod classify;
-pub(crate) mod composite;
 pub mod config;
 pub mod context_packager;
 pub mod contradiction;
